fix(status): write a status only while the row is still this server's
Review follow-ups. A manager listing an unclaimed row can be overtaken by
the manager that can see its pane claiming it mid-pass, so the status
write now carries the socket and lands only while the row is unclaimed or
this server's; a write that does not land announces nothing. The new
tests moved beside the files they cover.
